New header modifications and samples deletion
Modifying headers and adding it on the layout, in consequence deleting
it from `home.html` and adding it to `layout.html` because this will be
visible to almost any places on the portfolio website, but the animation
using it as a welcome screen will not.

`welcome_screen.html` is `header.html` right now.

Other files as `main.css` and `welcome_specials.css` was modified for a
better fitting on last modifications. The frontend will be improved this
afternoon to still working on backend.